-Used Flow instead of Coroutines to fetch both local and network data     -Used more of jetpack compose to display UI     -Updated PR template
