+++ app/src/main/java/com/example/newsly/database/TopStoryDao.kt
@@ -5,12 +5,13 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import com.example.newsly.model.TopStories
+import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface TopStoryDao {
 
     @Query("SELECT * from TopStories ORDER BY TopStories.id")
-    suspend fun getStories() : Array<TopStories>
+    fun getStories() : Flow<Array<TopStories>>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertStory(story : TopStories)
