+++ pull_request_template.md
@@ -1,1 +1,2 @@
-**Demo**
+**What Was Done**
+{latest_commit_message}
